Refactored Room Decoration Positioning
Previously, a users temporary changes were stored in groupRef, which was
kept track of using a furniturePositions dict and eventually synced to the
backend when the save changes button was clicked.
Now, the temporary changes are stored in roomSlice state, and sync
to backend when the button is clicked.
---
This change simplifies the code and makes it easier to manage the state
of the room decorations. It also allows for more flexibility in how the
decorations are positioned and updated.