Fixed Issue where purchased items were not synced
Fixed issue where purchased items were not synced with the backend.
The previous implementation had a bug where sometimes the purchased items
would not be correctly updated in the backend, leading to inconsistencies
when checking on your inventory. This commit fixes that issue by ensuring
that the syncing and fetching of inventory state is now synchronous within
one async thunk.
The issue with coins not updatung properly after habit completion seems to also
be resolved by this change.